Add ActivityPub ID to account model and to account crawling functions (#175)
I'm just gonna brazenly come in hot with a pull request right off the
bat. This one is a prerequisite for my approach to #160 (the rest will
be a breeze after this) and to doing much of any ActivityPub-compatible
stuff in the future, so I think it's worth it.

This is a small change that adds a column `activitypub_id` to the
account model. See https://www.w3.org/TR/activitypub/#obj-id for what
this is going to store. It also adjusts the account crawlers to populate
it for newly scanned accounts – we get the data for free with the
Mastodon API call that we already use.

I made the column `NULL` by default because we won't have this data for
existing accounts. It can't be reliably constructed offline from what we
already have. It can be backfilled with a WebFinger request for each
existing account: for `me@example.social`,

1. Fetch
`https://example.social/.well-known/webfinger?resource=acct:me@example.social`.
2. If successful, check for existence of the `links` array.
3. If present, filter for the first object with `type` ==
`application/activity+json`.
4. If found, save the `href` property of that object as the ActivityPub
ID for `me@example.social`.

However, these are network requests that should be politely spaced out,
given that the main fedidevs.com deployment probably has thousands of
accounts in the database. So we'd need some sort of maintenance task to
do this in the background, which goes past my Django chops. If I write
you a function that does the WebFinger stuff described above, would you
have a good idea where to put it?

And on a contribution process note, do feel free to tell me off if I'm
approaching this wrong or you would like more discussion in advance
rather than based on a pull request. I don't actually collaborate a ton
in the FOSS space so I might not be in tune with the cultural norms.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Anže Pečar <anze@pecar.me>