Add host-meta support to WebFinger lookup (#178)
This adds support for [RFC
6415](https://datatracker.ietf.org/doc/html/rfc6415) (“host-meta”) to
the WebFinger lookup command. This is needed for a few ActivityPub
servers in the wild and should resolve some missing ActivityPub account
IDs.

The implementation introduces a naive dict-based cache for WebFinger
URLs so the host-meta info only needs to be fetched once per instance
within one invocation of the command. I don't believe this information
is worth persisting.

As a side note, this PR also fixes an oversight in `crawlone.py` for
compatibility with older Mastodon versions.

Please note that my dev environment is/was not set up to really test
this properly. I hope it works in practice.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>