Add summary / alt text to starter pack splash image ActivityPub and rename starter pack test functions (#180)
A long title for two very small updates to the ActivityPub stuff, the
first of which takes way more lines of text to explain than to implement
and test:

It was bothering me that we had no facility to add alternative text to
our splash image in the ActivityPub representation. The ecosystem
generally uses the `summary` property to provide replacement text for
media, and that's inherited from `Object`, so we can use it for the
splash image without any trouble.

However, I have currently hard-coded the splash image alt text to be the
empty string. This is because I believe our splash images will be
“either decorative or supplemental to the rest of the content, redundant
with some other information in the document” ([WHATWG HTML Living
Standard](https://html.spec.whatwg.org/multipage/embedded-content.html#the-img-element:attr-img-alt-5)
on what it means when the alt attribute is the empty string). That is,
the splash image contains nothing that isn't already present in the
surrounding document, and serves only an aesthetic purpose. Having an
empty string as the alt text signifies this, and thus has a different
meaning than not having the alt text present at all. If the splash image
design becomes much richer than I am anticipating, we may need to
revisit this.

Point being, by adding an empty alt text / summary, we let consumers
know that the splash image is purely decorative, and nudge other future
implementers of the schema to think about whether their splash images
should perhaps have alternative text.

I have also explained this in a code comment and updated the
corresponding test alongside the implementation itself.

Speaking of the tests, the other change in this pull request is that I
renamed the API test functions. With respect, looking at them after the
last refactoring, I felt like their names no longer reflected their
purpose.

Assuming you don't have any philosophical disagreements, this should be
straightforward to merge.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>