Add splash image to starter pack model, update admin view, migrations, dependencies (#183)
This PR updates the starter pack model to contain a `splash_image`.
Using the `ImageField` type for this already introduces a dependency on
Pillow, meaning this lone commit does most of the plumbing that we're
going to need for splash images before I even start generating any. So I
thought I'd separate it out and send it in as is, since it already
touches a lot of places and I don't want my dev environment to diverge
too far from `main` for too long.

Next step will be to add a function to the starter pack view to render a
splash image and probably add an admin command to run it.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Anže Pečar <anze@pecar.me>