Add splash image rendering with admin and backend functions (#184)
Okay, here's the big one.

This PR adds splash image rendering for starter packs using Pillow.
Splash images are generated by fetching up to 24 avatars of accounts in
the starter pack, arranging them into a nice hex grid, adding the title
of the pack and saving everything into a local media directory.

Besides the splash image rendering itself, this PR contains an admin
function that allows you to (re-)render splash images for specific
starter packs in the admin view, as well as a backend command
`update_splash_images` that checks for starter packs needing splash
image updates and handles them in the background. It also contains data
model changes that mark starter packs where a splash image rerender
*would* be required (e.g. after adding or removing accounts in a way
that would influence the splash image content), but do not automatically
act on these markings.

This PR does **not** yet use the rendered starter packs in the UI or
triggers any automatic render processes outside of the two
administrative functions above (although the generated splash images are
technically web-accessible if you know the path). This version is
intended to test the feature's resource usage and resilience under real
circumstances.

This PR also does not contain any unit tests. Graphics rendering is
difficult to unit test (do we want to compare pixel color values?), and
outside of graphics, network, and storage output, you could at best test
details around the margins (e.g. "the number of avatars on a splash
image is <= the number of accounts in the starter pack) for which the
implementation is mostly trivial. Does this need to change, should we
have unit tests for splash images and if so, what should they look like?

I'm going to leave a few architectural concerns as review comments
myself. Open to any required changes as always.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>