Use starter pack splash images in metadata (OpenGraph and APIs) (#185)
This adds the splash image URL to the page metadata if a splash image
has been generated for the requested starter pack. The generic OpenGraph
image is replaced with the custom URL in the HTML and ActivityPub
versions. The JSON version gets the splash image URL only if one exists
(no fallback to the generic one). In effect, whenever this goes live,
the new splash images will show up when people share starter pack links.
The splash images are not yet used anywhere in the web UI itself, since
I'm not sure how (if at all) that's planned.

I had to get a little bit tricksy in the templates to make this work.
They used to assume that `og:image` and `twitter:image` would always
reside in the `static` directory. I have adjusted the templates so that
if `page_image` starts with a `/`, it is assumed to be a full path
(relative to the web root), and if it starts with anything else, it's a
static file. I hope that makes sense.

I've also rearranged the splash image rendering code a little bit to
nudge it towards being testable. It's now possible to render a splash
image without saving it to the file system. As a side effect of the
refactoring, this fixes a small bug where `splash_image` model entries
could be stored in the database without a corresponding image in the
file system if the `media` directory was non-writable at time of
execution.

This PR still includes no unit tests for the splash images because the
rendering function takes a starter pack as a parameter, extracts the
accounts, and then fetches the avatars from whatever URLs are in the
model. This process is kinda difficult to mock and will probably require
more refactoring to become suitable for unit testing. Still thinking
about it.

Just me being a bit impatient now. 🙂 I looked at a few splash image URLs
on the live site and didn't see any that were missing. If you think
we're ready to go with them, feel free to go ahead and merge, otherwise
let me know about any problems.

---------

Co-authored-by: Julian Fietkau <git@fietkau.software>